Update script.js
Step 14 Passed

It is time to start writing the script that makes your form work.

To access an HTML element with a given id name, you can use the getElementById() method. Here's an example of how to use this method:

<h1 id="title">Main title</h1>

const mainTitleElement = document.getElementById('title');

Begin by getting the form element (using the id) and storing it in a variable called calorieCounter.