Update script.js
Step 22

You will need to check if the array ["+", "-", " "] does not include the current character. You can use a combination of the includes() method and the ! operator to do this.

The .includes() method returns true if the array contains the character, and false if not. The logical NOT operator (!) will return the opposite of the value of the .includes() method.

Here is an example checking if the current character is not in the vowels array. If it is not, that current character is pushed into the consonantArray:

const inputString = "Hello World";
const charArray = inputString.split('');
const consonantArray = [];

for (let i = 0; i < charArray.length; i++) {
  if (!['a', 'e', 'i', 'o', 'u'].includes(charArray[i])) {
    consonantArray.push(charArray[i]);
  }
}

Within your loop, you need to check if the character in strArray at index i is not a "+", "-", or a space (" "). If it is not, push it to the cleanStrArray.