Update script.js
Step 31

The e in a number input can also be an uppercase E. Regex has a flag for this, however – the i flag, which stands for "insensitive".

/Hello/i

The following regex would match hello, Hello, HELLO, and even hElLo because of the i flag. This flag makes your pattern case-insensitive.

Add the i flag to your regex pattern.