Update script.js
Step 62

In programming, null is meant to represent the absence of a value. In this case, if the user enters an invalid input, you want to alert them and then return null to indicate that the function has failed.

Still within your if block, set isError to true and return null.