Update script.js
Step 81

When the user has a calorie deficit, the remainingCalories value will be negative. You don't want to display a negative number in the result string.

Math.abs() is a built-in JavaScript method that will return the absolute value of a number.

const num = -5;
Math.abs(num); // 5

In your span text, wrap your remainingCalories reference in Math.abs() to ensure that the value is positive.